requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
dj-database-url==1.0.0
Django==4.1.1
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
pycparser==2.21
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0
